LayoutTests/crypto/subtle/ec-import-spki-key-export-raw-key-expected.txt

Test importing a SPKI ECDH key and exporting it in raw format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS bytesToHexString(publicKey) is expectedRawKey
PASS successfullyParsed is true

TEST COMPLETE
